+++ frontend/src/index.js
@@ -10,7 +10,7 @@ import "bootstrap/dist/css/bootstrap.min.css";
 
 // Views
 import App from './App';
-import Home from './views/home';
+import Home from './views/home/home';
 import DetailedChecklist from './views/detailedChecklist';
 import Kanban from './views/kanban/kanban';
