+++ frontend/src/views/home/home.css
@@ -1,16 +1,15 @@
 @charset "UTF-8";
 @import url("https://fonts.googleapis.com/css2?family=Aref+Ruqaa:wght@400;700&display=swap");
+#root {
+  height: 100vh;
+}
+
 /*!
  * Bootstrap  v5.2.3 (https://getbootstrap.com/)
  * Copyright 2011-2022 The Bootstrap Authors
  * Copyright 2011-2022 Twitter, Inc.
  * Licensed under MIT (https://github.com/twbs/bootstrap/blob/main/LICENSE)
  */
-
-#root {
-  height: 100vh;
-}
-
 :root {
   --cc-blue: #0d6efd;
   --cc-indigo: #6610f2;
