+++ frontend/src/views/home/home.scss
@@ -28,4 +28,4 @@ $h6-font-size:                $font-size-base;
 
 $primary: #BF5676;
 
-@import "../../node_modules/bootstrap/scss/bootstrap";
+@import "../../../node_modules/bootstrap/scss/bootstrap";
